refactor: move Version to shared-internal
