+++ automation/code-generator/src/main/kotlin/io/github/typesafegithub/workflows/codegenerator/AddDeprecationInfo.kt
@@ -1,7 +1,7 @@
 package io.github.typesafegithub.workflows.codegenerator
 
 import io.github.typesafegithub.workflows.codegenerator.model.ActionBindingRequest
-import io.github.typesafegithub.workflows.codegenerator.model.Version
+import io.github.typesafegithub.workflows.shared.internal.model.Version
 
 fun List<ActionBindingRequest>.addDeprecationInfo(): List<ActionBindingRequest> =
     this.groupBy { "${it.actionCoords.owner}/${it.actionCoords.name}" }
